requirements.txt
streamlit>=1.31.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
aiohttp>=3.9.0
python-jose>=3.3.0